feat: complete implementation
