+++ packages/edge-gateway/package.json
@@ -22,7 +22,7 @@
     "ipfs-core-utils": "^0.15.0",
     "ipfs-gateway-race": "link:../ipfs-gateway-race",
     "itty-router": "^2.4.5",
-    "multiformats": "^9.6.4",
+    "multiformats": "^13.3.0",
     "p-any": "^4.0.0",
     "p-defer": "^4.0.0",
     "p-retry": "^5.0.0",
