+++ packages/edge-gateway/src/gateway.js
@@ -1,5 +1,5 @@
 /* eslint-env serviceworker, browser */
-/* global Response caches */
+/* global Response caches, IdentityTransformStream */
 
 import pAny, { AggregateError } from 'p-any'
 import pDefer from 'p-defer'
@@ -9,7 +9,8 @@ import { gatewayFetch } from 'ipfs-gateway-race'
 
 import {
   getCidFromSubdomainUrl,
-  toDenyListAnchor
+  toDenyListAnchor,
+  getCidFromEtag
 } from './utils/cid.js'
 import { getHeaders } from './utils/headers.js'
 import { getCidForbiddenResponse } from './utils/verification.js'
@@ -41,7 +42,7 @@ import {
  */
 
 /**
- * Handle gateway request.
+ * Handle gateway GET request.
  *
  * @param {Request} request
  * @param {Env} env
@@ -133,7 +134,7 @@ export async function gatewayGet (request, env, ctx) {
   } = await getFromGatewayRacer(cid, pathname, search, getHeaders(request), env, ctx)
 
   // Validation layer - resource CID
-  const resourceCid = pathname !== '/' ? getCidFromEtag(winnerGwResponse.headers.get('etag') || cid) : cid
+  const resourceCid = pathname !== '/' ? getCidFromEtag(winnerGwResponse.headers.get('etag') || `"${cid}"`) : cid
   if (winnerGwResponse && pathname !== '/' && resourceCid) {
     const resourceCidForbiddenResponse = await getCidForbiddenResponse(resourceCid, env)
     if (resourceCidForbiddenResponse) {
@@ -206,7 +207,7 @@ async function getFromCdn (request, env, cache) {
 /**
  * @param {Request} request
  * @param {Env} env
- * @param {string} cid
+ * @param {import('multiformats').UnknownLink} cid
  * @param {{ pathname?: string, search?: string }} [options]
  * @return {Promise<ProxiedCDNResponse | undefined>}
  */
@@ -276,7 +277,7 @@ async function getFromDotstorage (request, env, cid, options = {}) {
 
 /**
  *
- * @param {string} cid
+ * @param {import('multiformats').UnknownLink} cid
  * @param {string} pathname
  * @param {string} search
  * @param {Headers} headers
@@ -444,28 +445,6 @@ function getResponseWithCustomHeaders (
   }
 
   return clonedResponse
-}
-
-/**
- * Extracting resource CID from etag based on
- * https://github.com/ipfs/specs/blob/main/http-gateways/PATH_GATEWAY.md#etag-response-header
- *
- * @param {string} etag
- */
-function getCidFromEtag (etag) {
-  let resourceCid = decodeURIComponent(etag)
-
-  // Handle weak etag
-  resourceCid = resourceCid.replace('W/', '')
-  resourceCid = resourceCid.replaceAll('"', '')
-
-  // Handle directory index generated
-  if (etag.includes('DirIndex')) {
-    const split = resourceCid.split('-')
-    resourceCid = split[split.length - 1]
-  }
-
-  return resourceCid
 }
 
 /**
