+++ packages/ipfs-gateway-race/package.json
@@ -30,10 +30,10 @@
     "gateway"
   ],
   "dependencies": {
-    "@storacha/unixfs-dl": "^1.3.0",
+    "@storacha/unixfs-dl": "^1.4.0",
     "@web-std/fetch": "^4.1.0",
     "any-signal": "^3.0.1",
-    "multiformats": "^9.6.4",
+    "multiformats": "^13.3.0",
     "p-any": "^4.0.0",
     "p-map": "^5.3.0",
     "p-retry": "^5.0.0",
